refactor: use absolute imports
refactor: move style-data to constants folder
