+++ src/App.jsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import Header from "./components/Header/Header";
+import Header from "@/components/Header/Header";
